chore: update comments
